requirements.txt
APScheduler==3.11.2
certifi==2026.1.4
charset-normalizer==3.4.4
greenlet==3.3.1
idna==3.11
narwhals==2.16.0
numpy==2.4.2
packaging==26.0
pandas==3.0.1
plotly==6.5.2
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
